v1/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
# PostgreSQL (可选,如果使用PostgreSQL请取消注释)
# psycopg2-binary==2.9.9
# SQLite (默认,已包含在Python标准库中)

# 向量数据库 (二选一)
# Milvus (可选,需要Docker)
# pymilvus==2.3.4
# Chroma (推荐,轻量级)
chromadb==0.4.22
langchain==0.1.6
langchain-community==0.0.19
langchain-openai==0.0.5
langchain-anthropic==0.1.1
langgraph==0.0.20
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
httpx==0.26.0
pypdf==3.17.4
python-pptx==0.6.23
tiktoken==0.5.2
langsmith==0.0.83
redis==5.0.1
websockets==12.0
